新增系统接口和根绝角色ID查询对应的权限接口

+++ spring-cloud-provide/spring-cloud-mssp/src/main/java/org/com/zhump/mssp/service/IMsspPermissionService.java
@@ -17,7 +17,7 @@ public interface IMsspPermissionService {
 
     List<MsspPermission> selectByExample(MsspPermissionExample example);
 
-    MsspPermission selectByPrimaryKey(Integer id);
+    MsspPermission selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") MsspPermission record, @Param("example") MsspPermissionExample example);
 }
