+++ spring-cloud-provide/spring-cloud-mssp/src/main/java/org/com/zhump/mssp/service/impl/MsspSystemServiceImpl.java
@@ -37,7 +37,7 @@ public class MsspSystemServiceImpl implements IMsspSystemService {
     }
 
     @Override
-    public MsspSystem selectByPrimaryKey(Integer id) {
+    public MsspSystem selectByPrimaryKey(Long id) {
         return msspSystemMapper.selectByPrimaryKey(id);
     }
